Moved drawing code to assets, added sand texture.
