+++ wandering-warriors/calculator.kv
@@ -49,7 +49,14 @@
 <CuneiformDrawingInput>:
     canvas:
         Color:
-            rgb: 1, 1, 1
+            rgb: 1, 1, 1, 1
+        Rectangle:
+            source: 'assets/sandTexture.jpg'
+            size: self.size
+            pos: self.pos
+    canvas.before:
+        Color:
+            rgba: 0, 0, 0
         Rectangle:
             size: self.size
             pos: self.pos
